+++ Project1/models.py
@@ -16,22 +16,15 @@ class Users(db.Model):
 
     def __repr__(self):
         return self.email
+class Books(db.Model):
+    __tablename__="books"
+    isbn=db.Column(db.String,primary_key=True)
+    title=db.Column(db.String,nullable=False)
+    author=db.Column(db.String,nullable=False)
+    year=db.Column(db.String,nullable=False)
 
-
-class Book(db.Model):
-    __tablename__ = "books"
-    id = db.Column(db.Integer, primary_key=True)
-    isbn = db.Column(db.String(80), index=False, unique=True, nullable=False)
-    title = db.Column(db.String(80), index=True, unique=False, nullable=False)
-    author = db.Column(db.String(128))
-    year = db.Column(db.Integer, index=False, unique=False, nullable=False)
-
-    def __init__(self, isbn, title, author, year) :
+    def __init__(self,isbn,title,author,year):
         self.isbn = isbn
         self.title = title
         self.author = author
         self.year = year
-
-    def __repr__(self):
-        return self.title
-
