+++ Project1/templates/Register.html
@@ -50,14 +50,14 @@ header{
               <div class="input-group-prepend">
                 <span class="input-group-text" id="basic-addon1">Email ID</span>
               </div>
-              <input type="email" class="form-control" name="name" placeholder="Email" aria-label="Username" aria-describedby="basic-addon1">
+              <input type="email" class="form-control" name="name" placeholder="Email" aria-label="Username" aria-describedby="basic-addon1" required>
           </div>
           <br>
           <div class="input-group mb-3">
               <div class="input-group-prepend">
                 <span class="input-group-text" id="basic-addon1">Password</span>
               </div>
-              <input type="password" class="form-control" name="pwd" placeholder="Password" aria-label="Password" aria-describedby="basic-addon1" pattern=".{8,}" title="Eight or more characters">
+              <input type="password" class="form-control" name="pwd" placeholder="Password" aria-label="Password" aria-describedby="basic-addon1" pattern=".{8,}" title="Eight or more characters" required>
             </div>
             <br>
           <button type="submit" class="btn btn-secondary btn-lg">Register</button>
